Fixed label languages and other visual issues
The language of all the labels that were in spanish, were changed to english. As well as other visual changes such as: Hide the 'user' label from the header, and add a margin between the add button and the 'card and list' icons in the Organizations page, because in bigger monitors they overlap each other.